+++ NEWS.md
@@ -1,3 +1,9 @@
+# RTL 0.1.8
+
+## Updates and Fixes
+
++ `tradestats()`.
+
 # RTL 0.1.7
 
 ## New